SQL: use BIGINT where can be neccessary in parking_lots table

+++ sql_scripts/create_smartparking_parking_lots_table.sql
@@ -1,10 +1,10 @@
 CREATE table smartparking_parking_lots (
-    id INTEGER NOT NULL AUTO_INCREMENT,
+    id BIGINT(20) NOT NULL AUTO_INCREMENT,
     parking_lot_availability VARCHAR(10),
     address VARCHAR(255),
-    user_id LONG,
+    user_id BIGINT(20),
     latitude REAL,
     longitude REAL,
-    time_of_submission LONG,
+    time_of_submission BIGINT(20),
     PRIMARY KEY(id)
 );
